python/requirements.txt
websocket-client
Flask
python-dotenv
